+++ MotoShopRacing/EliminarProd.cs
@@ -74,7 +74,7 @@ namespace MotoShopRacing
 
         private void pictureBox1_Click(object sender, EventArgs e)
         {
-            Sql.Actualizar(int.Parse(ProdID), txtNombre.Text, txtMarca.Text, (int)numCant.Value, int.Parse(txtPrecio.Text));
+            Sql.Actualizar(int.Parse(ProdID), txtNombre.Text, txtMarca.Text, (int)numCant.Value, decimal.Parse(txtPrecio.Text));
             MessageBox.Show("Registro actualizado con exito!");
             dgvMantenimiento.DataSource = Sql.MostrarProductos();
         }
